% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW3LR.R
\name{print_regression_summary}
\alias{print_regression_summary}
\title{Print Regression Summary}
\usage{
print_regression_summary(results)
}
\arguments{
\item{results}{A list returned by the \code{linear_regression} function.}
}
\description{
This function takes the output from \code{linear_regression} and prints
all the important statistics and diagnostic information in a readable format.
}
